refactor: MapViewLayout 파일 생성 & 주석 처리
